requirements.txt
requests
google-search-results
aiohttp
PyPDF2
python-docx
openpyxl
python-pptx